back: Extract experience scope resolution into its own module.
Three handlers now ask the same question — may this caller act on
this experience, and which region should the audit row name — and
the answer lived inside curationController next to editExperience,
its first caller. Moving it to experienceScope.ts leaves that file
745 lines instead of 790 and puts the shared predicate where a
second and third caller can reach it without importing an editor.

No behaviour change: the query, the #450 fix it carries, and the
NULL-region rule for admins and category curators are unchanged.

[Issue: #480]
